fix: the FP64 WRMS was narrowed back to float at every consumption point
Codex stop-review: "FP64 WRMS result is narrowed back to FP32 before every gate decision."
Correct, and it made my previous commit's claim hollow: promoting the reduction while the value
lands in .item<float>() into float members, read into float locals at the gates, is a claim about
the MIDDLE of the pipe while the decision still reads float.

Widened end to end, each site found by following the value rather than the type names:

  * the members: last_stage_wrms_norm_ / last_stage_wrms_growth_ float -> double
  * the assignment: .item<float>() -> .item<double>() at the metric site
  * the gate locals: `const float growth = ...` and `float value = ...` -> double
  * the stage_gate_growth_value lambda: returned float, the last remaining narrowing
  * the MPI sync: the WRMS pair travelled as MPI_FLOAT, which would have narrowed ACROSS THE WIRE
    on multi-rank exactly what the promotion widened locally; those two now reduce as MPI_DOUBLE.
    The rel_R trio stays float deliberately -- widening it is a separate decision about a
    different metric, not something to sweep along.

The MPI leg was caught by the FULL build, not the CTest build -- sync_stage_quality_mpi is
MPI-gated code the test targets never compile, which is the known ./compile-vs-ctest coverage gap
doing its job in the right direction for once.

44/44 ctest. Force-cleaned rebuild: fingerprint byte-identical -- no decision flips on this
trajectory; what changes is that a large-domain or multi-rank run can no longer flip one through
float overflow, block loss, or wire narrowing.

+++ external/libtorch_wrf/sdirk3/wrf_sdirk3_tile_unified.h
@@ -1347,8 +1347,11 @@ private:
     float last_stage_rel_R_full_ = 0.0f;  // from RESIDUAL_REEVAL: floored ratio (K_floor for gate stability)
     float last_stage_rel_R_full_raw_ = 0.0f; // v20.14r41: raw ratio (no K_floor) for post-damp trigger
     float last_stage_R_full_norm_ = 0.0f; // v20.14r27v: absolute ||R_full|| for post-damp comparison
-    float last_stage_wrms_norm_ = 0.0f;   // P1: WRMS residual norm using stage state as reference
-    float last_stage_wrms_growth_ = 0.0f; // P1: WRMS(R_k) / WRMS(R_0), production stage-gate metric
+    // DOUBLE end to end. The reduction inside wrms_norm_packed runs in FP64; narrowing the
+    // result to float here -- which these were -- handed the gate an FP32 number anyway, making
+    // the promotion a claim about the middle of the pipe while the DECISION still read float.
+    double last_stage_wrms_norm_ = 0.0;   // P1: WRMS residual norm using stage state as reference
+    double last_stage_wrms_growth_ = 0.0; // P1: WRMS(R_k) / WRMS(R_0), production stage-gate metric
     float last_stage_initial_R0_ = 0.0f;  // v20.14r39: Newton's initial unscaled L2 ||R_0|| for diagnostics
     int progress_low_streak_ = 0;         // v20.14r64: consecutive low-progress step counter
